+++ libdraw/rgb.c
@@ -47,7 +47,7 @@ int
 rgb2cmap(int cr, int cg, int cb)
 {
 	int i, r, g, b, sq;
-	ulong rgb;
+	uint32 rgb;
 	int best, bestsq;
 
 	best = 0;
